+++ affinidi-did-resolver-cache-sdk/src/config.rs
@@ -1,17 +1,17 @@
 //! Handles the initial configuration for the DID Cache Client.
 //!
-//! Call the [ClientConfigBuilder] to create a new configuration.
+//! Call the [DIDCacheConfigBuilder] to create a new configuration.
 //!
 //! Example: Running in local mode with defaults:
 //! ```rust
-//! use affinidi_did_resolver_cache_sdk::config::ClientConfigBuilder;
-//! let config = ClientConfigBuilder::default().build();
+//! use affinidi_did_resolver_cache_sdk::config::DIDCacheConfigBuilder;
+//! let config = DIDCacheConfigBuilder::default().build();
 //! ```
 //!
 //! Example: Running in network mode with custom settings:
 //! ```rust
-//! use affinidi_did_resolver_cache_sdk::config::ClientConfigBuilder;
-//! let config = ClientConfigBuilder::default()
+//! use affinidi_did_resolver_cache_sdk::config::DIDCacheConfigBuilder;
+//! let config = DIDCacheConfigBuilder::default()
 //!     .with_network_mode("ws://127.0.0.1:8080/did/v1/ws")
 //!     .with_cache_capacity(500)
 //!     .with_cache_ttl(60)
@@ -25,12 +25,12 @@
 use std::time::Duration;
 use wasm_bindgen::prelude::*;
 
-/// Private Configuration for the client.
+/// Configuration for the DID Cache client.
 ///
-/// Use the [ClientConfigBuilder] to create a new configuration.
+/// Use the [DIDCacheConfigBuilder] to create a new configuration.
 #[derive(Clone, Debug)]
 #[wasm_bindgen(getter_with_clone)]
-pub struct ClientConfig {
+pub struct DIDCacheConfig {
     #[cfg(feature = "network")]
     pub(crate) service_address: Option<String>,
     pub(crate) cache_capacity: u32,
@@ -43,7 +43,7 @@ pub struct ClientConfig {
     pub(crate) max_did_size_in_kb: f64,
 }
 
-/// Config Builder to construct options required for the client.
+/// DID Cache Config Builder to construct options required for the client.
 /// You must at least set the service address.
 ///
 /// - service_address: REQUIRED: The address of the service to connect to.
@@ -51,7 +51,7 @@ pub struct ClientConfig {
 /// - cache_ttl: The time-to-live in seconds for each item in the local cache (default: 300 (5 Minutes)).
 /// - network_timeout: The timeout for network requests in milliseconds (default: 5000 (5 seconds)).
 /// - network_cache_limit_count: The maximum number of items to store in the network cache (default: 100).
-pub struct ClientConfigBuilder {
+pub struct DIDCacheConfigBuilder {
     #[cfg(feature = "network")]
     service_address: Option<String>,
     cache_capacity: u32,
@@ -64,7 +64,7 @@ pub struct ClientConfigBuilder {
     max_did_size_in_kb: f64,
 }
 
-impl Default for ClientConfigBuilder {
+impl Default for DIDCacheConfigBuilder {
     fn default() -> Self {
         Self {
             #[cfg(feature = "network")]
@@ -81,7 +81,7 @@ impl Default for ClientConfigBuilder {
     }
 }
 
-impl ClientConfigBuilder {
+impl DIDCacheConfigBuilder {
     /// Enables network mode and sets the service address.
     /// Example: `ws://127.0.0.1:8080/did/v1/ws`
     #[cfg(feature = "network")]
@@ -135,8 +135,8 @@ impl ClientConfigBuilder {
     }
 
     /// Build the [ClientConfig].
-    pub fn build(self) -> ClientConfig {
-        ClientConfig {
+    pub fn build(self) -> DIDCacheConfig {
+        DIDCacheConfig {
             #[cfg(feature = "network")]
             service_address: self.service_address,
             cache_capacity: self.cache_capacity,